chatbot/requirements.txt
langchain
langchain_community
langchain-google-genai
pymongo
pypdf
bs4
requests
python-dotenv
langchain-mongodb
Flask
flask-cors
Gunicorn
langchain_pinecone
pinecone-notebooks
uuid
sentence_transformers